Variable Input
I made a linear control for the variable input. Yet, no matter what (that is, I added time delay, used sample and hold, used the output voltage as reference to hold it ), the duty cycle change can not catch the output's fast response. Hence, the ripple shows a sinusoidal form and the ripple exceeds the desired value.