Add files via upload
Since the diode-capacitor has not been working for the controlled Cuk converter, it has been removed. Instead, a single diode has been added just before the load resistance in order to eliminate the positive part of the oscillation on the output end of the load resistor. Previously, there was an oscillation between  +5 volts and -25 volts on V(out). With that, the oscillation is now between 0 volts and -25 volts. The result is not still desirable but it is better. Soon, V(out) value will be nearly -15 volts DC, which is the desired potential difference.