Add files via upload
In the uncontrolled part of Cuk converter, the previous output was about -17 volts DC. In this version, it has been fixed by putting a diode whose forward voltage is 2 volts just after the input end of the converter. And now, the output voltage which we obtain is -15 volts DC with almost no ripple.