Add files via upload
An error in the gate controller of full bridge converter has been fixed, an LC filter and a 15 volts forward biased diode has been added, which are parallel to each other. The function of LC filter is to eliminate the AC component of the incoming signal and the function of 15 volts forward biased diode is to fix the output voltage to 15 volts.